64bit

+++ include/kernel/atag.h
@@ -33,6 +33,6 @@ typedef struct atag {
     };
 } atag_t;
 
-uint32_t get_mem_size(atag_t * atags);
+uint64_t get_mem_size(atag_t * atags);
 
 #endif
